Streamlit_Version/requirements.txt
streamlit
python-dotenv
tafrigh
anthropic
ffmpeg
tafrigh[wit]